[require migrate:refresh --seed] update CategorySeeder to get taxonomy from MSMS and insert to DB

+++ database/seeds/CategorySeeder.php
@@ -5,6 +5,7 @@ use App\Beacon;
 use App\BeaconMinor;
 use App\Category;
 use App\Item;
+use App\Repositories\CategoryRepository;
 use App\WatchingList;
 use Illuminate\Database\Seeder;
 use Illuminate\Database\Eloquent\Model;
@@ -14,12 +15,26 @@ class CategorySeeder extends Seeder {
 	public function run()
 	{
         DB::table('category_minor')->delete();
-        DB::table('categories')->delete();
 
-        $catFabricSofteners=Category::create(['id' => '1115193_1071967_1149392']);
-        $catLaundryDetergents=Category::create(['id' => '1115193_1071967_1149379']);
-        $catToothpaste=Category::create(['id' => '1085666_1007221_1023020']);
-        $catSoftDrinks=Category::create(['id' => '976759_976782_1001680']);
+        $forceRefresh=false;
+        if (!$forceRefresh&&!Category::all()->isEmpty()){
+            $this->command->info('Table categories not empty, skipped this table. Turn on forceRefresh if you want!');
+            return ;
+        }
+
+        DB::table('categories')->delete();
+        $catRepo=App::make('App\Repositories\CategoryInterface');
+        $taxonomy = $catRepo->getTaxonomy(true);
+        $categories = $taxonomy['categories'];
+        foreach ($categories as $category) {
+            $this->insert($category,null);
+        }
+
+
+        $catFabricSofteners=Category::find('1115193_1071967_1149392');
+        $catLaundryDetergents=Category::find('1115193_1071967_1149379');
+        $catToothpaste=Category::find('1085666_1007221_1023020');
+        $catSoftDrinks=Category::find('976759_976782_1001680');
 
         $minor1=BeaconMinor::find(1);
         $minor2=BeaconMinor::find(2);
@@ -30,5 +45,21 @@ class CategorySeeder extends Seeder {
         $minor2->categories()->attach($catLaundryDetergents);
         $minor3->categories()->attach([$catSoftDrinks->id, $catToothpaste->id]);
 	}
+
+    private function insert(array $category, $parent)
+    {
+        $cat = new Category(['id' => $category['id'], 'name' => $category['name'], 'parent_id' => $parent['id']]);
+        if (array_key_exists('children', $category)) {
+            $cat->is_leaf=false;
+            $cat->save();
+            foreach ($category['children'] as $child) {
+                $this->insert($child,$category);
+            }
+        }
+        else{
+            $cat->is_leaf=true;
+            $cat->save();
+        }
+    }
 
 }
